Fix compilation error on Windows for context error handler
This fixes a compilation error which occurs on Windows when
registering the default error handler callback when creating
a new context object.

In OpenCL 1.1 and later the callback function is expected to
use the __stdcall calling convention. This is optionally defined
by the CL_CALLBACK macro on WIN32 platforms. If available, it is
defined with the BOOST_COMPUTE_CL_CALLBACK macro which is then
used to annotate the callback functions.