Fix partition_by_counts test for Intel CPUs
This fixes a test failure in partition_by_counts for Intel CPU
devices. The clCreateSubDevices() function is not guaranteed to
return the sub-devices in the same order as they were specified
in the compute unit counts argument.

Now the test sorts the returned sub-devices by number of compute
units and checks that each specified count is present.

See issue #185.