Fix kernels in reduce_by_key_with_scane.hpp
It was not guaranteed that every barrier would be encountered by all
work-items in a work-group executing the kernels due to some work-items
returing too early, not entering the conditional if with with a barrier
or not executing every iteration of the loop with a barrier.

See barrier function description in OpenCL documentation.