Work around libc++ bug 17782
libc++ fails to extract a valid float/double literal when it is followed
by any special characters that can also be found in a float value,
in this case, f (https://llvm.org/bugs/show_bug.cgi?id=17782).

This commit works around the problem by not putting the "f" suffix into
the string for extraction.